Move list page to /bowls and keep /user redirect compatibility
